breeding/pyproject.toml
[project]
name = "fractal-breeding"
version = "0.1.0"
description = "Latent space breeding and inference server for fractal flame GAN"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.4",
    "pillow>=10.0",
    "numpy>=1.26",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "pydantic>=2.0",
]

[tool.setuptools.packages.find]
include = ["breeding*"]
